Refactor clustering into a new module with a 3D-aware interface
Extract the monolithic clusters.py (~1900 lines) into a new
thor/clustering/ package with focused modules:

- algorithms.py: ClusteringAlgorithm protocol operating on
  TransformedDetections (supports future 3D algorithms)
- velocity_grid.py: VelocityGridBase ABC with shared velocity-grid
  sweep logic and covariance-aware parameter calculation
- dbscan.py: Thin VelocityGridDBSCAN subclass
- hotspot2d.py: Numba functions + thin VelocityGridHotspot2D subclass
- data.py: Clusters/ClusterMembers/FittedClusters/FittedClusterMembers
- fitting.py: Polynomial motion model fitting
- metrics.py: Cluster filtering by arc length, nights, etc.

clusters.py is now a backward-compatible re-export shim. main.py
instantiates algorithm classes directly from config. config.py widens
cluster_algorithm from Literal to str for extensibility.
