+++ src/thor/__init__.py
@@ -3,7 +3,7 @@ try:
     from ._version import __version__
 except ImportError:
     __version__ = "unknown"
-from .clusters import *
+from .clustering import *
 from .config import *
 from .main import *
 from .orbit import *
